+++ src/test/java/ru/skypro/homework/controller/TestAddControllerWithAnonymousUser.java
@@ -10,12 +10,11 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.test.context.support.WithAnonymousUser;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import ru.skypro.homework.models.mappers.*;
@@ -33,24 +32,21 @@ import java.util.Optional;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static ru.skypro.homework.controller.Constants.*;
 
 @WebMvcTest(controllers = AdsController.class)
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration
-public class TestAddControllerPost {
+@WithAnonymousUser
+public class TestAddControllerWithAnonymousUser {
     private final String LOCAL_URL = URL + PORT + "/" + "ads";
     private final AdsMapper adsMapperTest = new AdsMapperImpl();
     private final CommentsMapper commentsMapperTest = new CommentsMapperImpl();
-    private final ImagesMapper imagesMapperTest = new ImagesMapperImpl();
     private final UserMapper userMapperTest = new UserMapperImpl();
     private final JSONObject adsCommentDto = new JSONObject();
-
     private final JSONObject adsJsonDTO = new JSONObject();
 
     @Autowired
@@ -83,6 +79,8 @@ public class TestAddControllerPost {
     private ImagesMapper imagesMapper;
     @InjectMocks
     private AdsController adsController;
+    @MockBean
+    private Authentication auth;
 
     @Before
     public void startData() {
@@ -113,7 +111,6 @@ public class TestAddControllerPost {
         adsJsonDTO.put("price", "105");
         adsJsonDTO.put("title", "ads title");
 
-        mockMvc = MockMvcBuilders.webAppContextSetup(context).build();
     }
 
     @Test
@@ -121,78 +118,61 @@ public class TestAddControllerPost {
         assertThat(adsController).isNotNull();
     }
 
-
     @Test
-    @WithMockUser(username = AUTHORS_EMAIL, value = AUTHORS_EMAIL)
-    public void testPostAdsComments() throws Exception {
-        when(adsRepository.findById(ADS_ID)).thenReturn(Optional.of(ADS_MODEL));
-        when(commentsMapper.toComments(any())).thenReturn(NEW_COMMENTS_MODEL);
-        when(adsCommentsRepository.save(any())).thenReturn(NEW_COMMENTS_MODEL);
+    public void testGetAllAdsWithoutAuthority() throws Exception {
+        mockMvc = MockMvcBuilders.webAppContextSetup(context).build();
 
-        when(userRepository.findUserByEmail(any())).thenReturn(Optional.of(AUTHOR_MODEL));
-        when(commentsMapper.toCommentsDto(any())).thenReturn(COMMENTS_DTO);
-
-        mockMvc.perform((post(LOCAL_URL + "/" + ADS_ID + "/comments")
-                        .content(adsCommentDto.toString())
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)))
+        when(adsRepository.findAll()).thenReturn(LIST_ADS);
+        when(auth.isAuthenticated()).thenReturn(false);
+        mockMvc.perform(get(LOCAL_URL))
                 .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.author").value(AUTHORS_ID))
-                .andExpect(jsonPath("$.createdAt").value(COMMENTS_DTO.getCreatedAt()))
-                .andExpect(jsonPath("$.pk").value(COMMENT_ID))
-                .andExpect(jsonPath("$.text").value(TEXT));
-
+                .andExpect(status().isOk());
     }
 
     @Test
-    @WithMockUser(username = AUTHORS_EMAIL, value = AUTHORS_EMAIL)
-    public void testPatchAdsComments() throws Exception {
+    public void testGetOneAdsWithoutAuthority() throws Exception {
         when(adsRepository.findById(ADS_ID)).thenReturn(Optional.of(ADS_MODEL));
-        when(commentsMapper.toComments(any())).thenReturn(NEW_COMMENTS_MODEL);
-        when(adsCommentsRepository.save(any())).thenReturn(NEW_COMMENTS_MODEL);
+        when(adsMapper.toFullAdsDto(ADS_MODEL)).thenReturn(FULL_ADS_DTO);
+        mockMvc.perform(get(LOCAL_URL + "/" + ADS_ID)
+                        .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andExpect(status().isUnauthorized());
+    }
+
+    @Test
+    public void testGetCommentsOneAdWithoutAuthority() throws Exception {
+        when((adsRepository.findById(ADS_ID))).thenReturn(Optional.of(ADS_MODEL));
+        when((adsCommentsRepository.findCommentsByAds(ADS_MODEL))).thenReturn(MODEL_COMMENT_LIST);
+        when(commentsMapper.toCommentsDto(any())).thenReturn(COMMENTS_DTO);
+        mockMvc.perform(get(LOCAL_URL + "/" + ADS_ID + "/comments/")
+                        .param("ad_pk", String.valueOf(ADS_ID))
+                        .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andExpect(status().isUnauthorized());
+    }
+
+    @Test
+    public void testGetAdsCommentsWithoutAuthority() throws Exception {
+        when((adsRepository.findById(ADS_ID))).thenReturn(Optional.of(ADS_MODEL));
         when(adsCommentsRepository.findById(COMMENT_ID)).thenReturn(Optional.of(COMMENTS_MODEL));
+        when(commentsMapper.toCommentsDto(COMMENTS_MODEL)).thenReturn(COMMENTS_DTO);
 
-        when(userRepository.findUserByEmail(any())).thenReturn(Optional.of(AUTHOR_MODEL));
-        when(commentsMapper.toCommentsDto(any())).thenReturn(COMMENTS_DTO);
-
-        mockMvc.perform((patch(LOCAL_URL + "/" + ADS_ID + "/comments/" + COMMENT_ID)
-                        .content(adsCommentDto.toString())
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.author").value(AUTHORS_ID))
-                .andExpect(jsonPath("$.createdAt").value(COMMENTS_DTO.getCreatedAt()))
-                .andExpect(jsonPath("$.pk").value(COMMENT_ID))
-                .andExpect(jsonPath("$.text").value(TEXT));
+        mockMvc.perform(get(LOCAL_URL + "/" + ADS_ID + "/comments/" + COMMENT_ID)
+                        .param("ad_pk", String.valueOf(ADS_ID))
+                        .param("id", String.valueOf(COMMENT_ID))
+                        .accept(MediaType.APPLICATION_JSON_VALUE))
+                .andExpect(status().isUnauthorized())
+        ;
     }
 
     @Test
-    @WithMockUser(username = AUTHORS_EMAIL, value = AUTHORS_EMAIL)
-    public void testPostAds() throws Exception {
-        when(imageRepository.save(any())).thenReturn(IMAGE);
-        when(userRepository.findUserByEmail(AUTHORS_EMAIL)).thenReturn(Optional.of(AUTHOR_MODEL));
-        when(adsMapper.fromCreateAds(any(), any(), any())).thenReturn(ADS_MODEL);
-        when(adsRepository.save(ADS_MODEL)).thenReturn(ADS_MODEL);
-        when(adsMapper.toAdsDto(ADS_MODEL)).thenReturn(adsMapperTest.toAdsDto(ADS_MODEL));
-        MockMultipartFile image = new MockMultipartFile("image", "filename.jpeg", "image/jpeg", "image.png".getBytes());
-
-        mockMvc.perform(MockMvcRequestBuilders.multipart(LOCAL_URL)
-                        .file(image)
-                        .param("properties", String.valueOf(adsJsonDTO))
-                        .content(adsJsonDTO.toString())
-                        .accept(MediaType.APPLICATION_JSON))
+    public void testGetMyAdsWithoutAuthority() throws Exception {
+        when(userRepository.findUserByEmail(any())).thenReturn(Optional.of(AUTHOR_MODEL));
+        when(adsRepository.findAll()).thenReturn(LIST_ADS);
+        mockMvc.perform(get(LOCAL_URL + "/me"))
                 .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.author").value(AUTHORS_ID))
-                .andExpect(jsonPath("$.image").value(IMAGE_ID))
-                .andExpect(jsonPath("$.pk").value(ADS_ID))
-                .andExpect(jsonPath("$.price").value(PRICE))
-                .andExpect(jsonPath("$.title").value(TITLE));
-
-
+                .andExpect(status().isUnauthorized());
     }
+
+
 
 
 }
