--- src/test/java/ru/skypro/homework/controller/WithCustomUser.java
@@ -1,15 +1,0 @@
-package ru.skypro.homework.controller;
-
-import org.springframework.security.test.context.support.WithSecurityContext;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-import static ru.skypro.homework.controller.Constants.AUTHORS_EMAIL;
-
-@Retention(RetentionPolicy.RUNTIME)
-@WithSecurityContext(factory = CustomSecurityContextFactory.class)
-public @interface WithCustomUser {
-
-    String username() default AUTHORS_EMAIL;
-}
